[Refact] #82 - BindingAdapter 코드 위치 조정
