+++ app/src/main/java/org/cazait/ui/adapter/BindingAdapter.kt
@@ -2,9 +2,12 @@ package org.cazait.ui.adapter
 
 import android.view.View
 import android.widget.ImageView
+import android.widget.TextView
 import androidx.core.view.isVisible
 import androidx.databinding.BindingAdapter
 import coil.load
+import org.cazait.R
+import org.cazait.model.CafeStatus
 
 const val DURATION_CROSS_FADE = 300
 
@@ -18,5 +21,18 @@ fun ImageView.setImage(imageUrl: String?) {
     // `crossfade`으로 사용자 경험 개선
     load(imageUrl) {
         crossfade(DURATION_CROSS_FADE)
+    }
+}
+
+@BindingAdapter("status")
+fun TextView.setStatus(cafeStatus: CafeStatus?) {
+    text = when (cafeStatus) {
+        null -> context.getString(R.string.state_normal)
+        CafeStatus.FREE -> context.getString(R.string.state_free)
+        CafeStatus.NORMAL -> context.getString(R.string.state_normal)
+        CafeStatus.CLOSE -> context.getString(R.string.state_close)
+        CafeStatus.CROWDED -> context.getString(R.string.state_crowded)
+        CafeStatus.VERY_CROWDED -> context.getString(R.string.state_very_crowded)
+        CafeStatus.NONE -> context.getString(R.string.state_normal)
     }
 }
